+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Smartmi fans will now log the total use time in minutes when connected to the device
+
 ### Changed
 - Added back the rotation direction service, the switch can now be used on Smartmi fans to switch between buzzer sound level (loud or quiet)
 - Updated README